SERVERLESS-354 Implement log shipping inside the go-proxy
This is a first implementation of 3rd party log-shipping inside the go-proxy. If the respective env vars are set, stdout and stderr from the function process are intercepted and written to the service as configured.

This contains initial support for DataDog, Papertrail and Logtail. Only HTTPS transport is currently supported.