Log the sentinels in the child process rather than the proxy process
The child process (aka the process executing the function) is communicating through 4 pipes with the outside world:

- stdin for getting work
- fd3 for writing responses out
- stdout and stderr for writing log output

The only synchronization happening between those pipes is the go-proxy writing into stdin and then waiting for a response to arrive at fd3. However, the arrival of the result is no guarantee that the log pipes (stdout and stderr) are consistent in that all the data in there has been flushed. Therefore, writing the sentinels inside of the go-proxy is inherently racing what's happening inside the respective function.

To fix this, we should change all function runtime implementations to write the log sentinels after writing the result to fd3 but before reading the next request. We can then guarantee that when we see these sentinels, all log lines have been dealt with.