SERVERLESS-961 Implement batching for the HTTP logger
This implements a fairly simple batching mechanism into the HTTP logger. Batches are written every N milliseconds (kicked off by the first entry of a batch) or if a batch exceeds a defined limit or if it is forced through a call to Flush (which we always do at the end of an activation).