SERVERLESS-958 Support multiple log destinations at once
This implements parity with APs ability to configure more than one log destination for a component. I specifically opted to not do the requests in parallel (yet) as it'd add quite a bit of complexity (handling errors, ideally pre-launching a pool of goroutines rather than launching one per request etc.) that I'd like to discuss separately. I'd also like to check how common that case even is and if it warrants the extra complexity.