Add the possibility to write Golang function as Golang-native HTTP handlers
This enables functions to take the form `func(http.ResponseWriter, *http.Request)` as is the Golang-native way of writing HTTP handlers in the standard library. It does that by translating a request as produced by OpenWhisk into a `*http.Request` and recording the respective response in-memory. It too will then be translated into a format OpenWhisk understand and is then returned.