Propagate __OW_DEADLINE into the context of a function
Now that we allow users to set a context.Context parameter in their function signature, we can take advantage of that and actually enrich its experience by surfacing the deadline we have for a function invocation on the context itself. That'll do 2 things:

1. The ctx.Deadline() function will return the respective deadline as a strongly typed time.Time.
2. The context itself will cancel once the deadline hits, so it can be used when calling other functions accepting contexts (like HTTP requests) to automatically cancel them as soon as the function deadline hits.

Note: For the second use-case, it might be too tight a timeframe to actually do something useful with the cancelation signal. We should discuss either adding some leeway (i.e. making the deadline a bit shorter by default in the function or a bit longer in the invoker) or document an easy way to grab the context's deadline and construct your own with a shorter deadline.