Stop blindly capitalizing 'main' (#21)
* Stop blindly capitalizing 'main'

'main' defines which function the Golang runtime executes on an invocation. The go-proxy defaults 'main' to "main", which usually doesn't work for Golang functions, since that'd collide with the "special" main function in Golang. So we called 'capitalize' to fix that (it turns "main" into "Main").

However, it destroys all kinds of other usages. If one defines "WebAction", it'll transform that into "Webaction", which is wrong.

As such, we want to only specifically fix the main -> Main conversion and leave the parameter alone for all other cases.

* Adjust tests to conform to new behavior