SERVERLESS-1376 Allow prelaunching binaries on startup (#25)
This adds a new mode to the go-proxy which allows it to launch a child binary on startup to amortize the cost of doing so. In that mode, which is triggered by the presence of an environment variable, the child process is expected to be able to deal with initialization itself. As such, it's passed the raw init request and is expected to respond with the "usual" OK response.