Always intercept logs even if not remote logging
Currently, we only intercept the log channels of the runtime process in the goproxy if we're using remote logging. That means that if we want to make use of remote logging, other features might not be usable, like runtime prelaunching.

This changes the behavior to ALWAYS intercept the log pipes and inserts a default logger that forwards to stdout and stderr, as today.

In doing so, this also fixes a bug where, when we've been using remote logging and the runtime would not write a sentinel, the process would timeout.