Refactor, change to snake_case, backport things.
Refactors the context building function so that all the context building happens in it, deadline and data from env vars. Switches the casing used for the context values from this-case to snake_case. Backports the behavior in our Go 1.17 runtime that allowed a variety of function signatures (including a few that used context.Context) to Go 1.15. In order to backport that, backported a function from the Go 1.17 stdlib to the code for our Go 1.15 runtime by including it inline in the file.